The no_pumping_n_wwtps.py has been modified to incorporate comments and other minor improvements.
THe operwa_library.py script is under modification to include Odur's contribution, which is under the temporal folder \Operwas_pump\src\faster\optimization\Odur, which needs to be removed once we finish the merge.

Some plot scripts to read the optimisation results are included. The file to run is plot_leo.py, which calls the funcitons in functions_plot_leo.py.